cleaning the route from util functions
